[#29] refactor: todoListBox 옵션, 정렬메서드 분리
