+++ src/components/TodoList/TodoList.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import styled from 'styled-components/macro';
 import { useTodosState } from 'context/todoContext/TodoContext';
 import TodoListBox from 'components/TodoListBox/TodoListBox';
+import { TStatus } from 'types';
 
 interface TodoListProps {}
 
@@ -12,9 +13,9 @@ const TodoList: React.FC<TodoListProps> = (props: TodoListProps) => {
   const completed = todos.filter((todo) => todo.status === '완료');
   return (
     <TodoListContainer>
-      <TodoListBox title="Not Started" todos={notStarted} />
-      <TodoListBox title="In Progress" todos={onGoing} />
-      <TodoListBox title="Completed" todos={completed} isLast />
+      <TodoListBox title="Not Started" status={TStatus.NOT_STARTED} todos={notStarted} />
+      <TodoListBox title="In Progress" status={TStatus.ONGOING} todos={onGoing} />
+      <TodoListBox title="Completed" status={TStatus.FINISHED} todos={completed} isLast />
     </TodoListContainer>
   );
 };
