+++ src/components/TodoListBox/data/dropDownOptions.ts
@@ -1,0 +1,24 @@
+const dropDownOptions = [
+  {
+    print: '사용자지정',
+    data: 'default',
+  },
+  {
+    print: '신규순',
+    data: 'latest',
+  },
+  {
+    print: '오래된순',
+    data: 'oldest',
+  },
+  {
+    print: '중요도',
+    data: 'priority',
+  },
+  {
+    print: '마감임박',
+    data: 'due',
+  },
+];
+
+export { dropDownOptions };
